+++ hyperbench/tests/utils/sparse_utils_test.py
@@ -7,17 +7,19 @@ from hyperbench.utils import sparse_dropout
 
 @pytest.fixture
 def mock_indices():
-    return torch.tensor([[0, 1, 2], [0, 1, 2]])
+    return torch.tensor([[0, 1, 2], [0, 1, 2]], dtype=torch.long)
 
 
 @pytest.fixture
 def mock_values():
-    return torch.tensor([1.0, 2.0, 3.0])
+    return torch.tensor([1.0, 2.0, 3.0], dtype=torch.float)
 
 
 def test_dropout_zero_probability(mock_indices, mock_values):
     """Test that zero dropout probability returns the original sparse tensor."""
-    sparse_tensor = torch.sparse_coo_tensor(mock_indices, mock_values, (3, 3))
+    sparse_tensor = torch.sparse_coo_tensor(
+        mock_indices, mock_values, (3, 3), dtype=mock_values.dtype
+    )
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.0)
 
@@ -27,18 +29,24 @@ def test_dropout_zero_probability(mock_indices, mock_values):
 
 def test_dropout_full_probability(mock_indices, mock_values):
     """Test that full dropout probability (1.0) drops all elements."""
-    sparse_tensor = torch.sparse_coo_tensor(mock_indices, mock_values, (3, 3))
+    sparse_tensor = torch.sparse_coo_tensor(
+        mock_indices, mock_values, (3, 3), dtype=mock_values.dtype
+    )
 
     result = sparse_dropout(sparse_tensor, dropout_prob=1.0)
 
     # All values should be zero when fill_value is 0
-    assert torch.allclose(result.coalesce().values(), torch.zeros_like(mock_values))
+    assert torch.allclose(
+        result.coalesce().values(), torch.zeros_like(mock_values, dtype=mock_values.dtype)
+    )
 
 
 @pytest.mark.parametrize("invalid_prob", [-0.5, 1.5])
 def test_dropout_invalid_probability_out_of_range(mock_indices, mock_values, invalid_prob):
     """Test that dropout probability below 0 raises ValueError."""
-    sparse_tensor = torch.sparse_coo_tensor(mock_indices, mock_values, (2, 2))
+    sparse_tensor = torch.sparse_coo_tensor(
+        mock_indices, mock_values, (2, 2), dtype=mock_values.dtype
+    )
 
     with pytest.raises(
         ValueError,
@@ -49,9 +57,9 @@ def test_dropout_invalid_probability_out_of_range(mock_indices, mock_values, inv
 
 def test_dropout_preserves_indices():
     """Test that dropout preserves the sparsity pattern (indices) unchanged."""
-    indices = torch.tensor([[0, 1, 2, 0], [0, 1, 2, 2]])
-    values = torch.tensor([1.0, 2.0, 3.0, 4.0])
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (3, 3))
+    indices = torch.tensor([[0, 1, 2, 0], [0, 1, 2, 2]], dtype=torch.long)
+    values = torch.tensor([1.0, 2.0, 3.0, 4.0], dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(indices, values, (3, 3), dtype=values.dtype)
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5)
 
@@ -62,9 +70,9 @@ def test_dropout_preserves_indices():
 def test_dropout_preserves_shape():
     """Test that dropout preserves the tensor shape."""
     shape = (5, 10)  # Shape of the tensor if it were dense
-    indices = torch.tensor([[0, 2, 4], [1, 5, 9]])
-    values = torch.tensor([1.0, 2.0, 3.0])
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, shape)
+    indices = torch.tensor([[0, 2, 4], [1, 5, 9]], dtype=torch.long)
+    values = torch.tensor([1.0, 2.0, 3.0], dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(indices, values, shape, dtype=values.dtype)
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5)
 
@@ -73,7 +81,7 @@ def test_dropout_preserves_shape():
 
 def test_dropout_preserves_dtype():
     """Test that dropout preserves the tensor dtype."""
-    indices = torch.tensor([[0, 1], [0, 1]])
+    indices = torch.tensor([[0, 1], [0, 1]], dtype=torch.long)
     values = torch.tensor([1.0, 2.0], dtype=torch.float32)
     sparse_tensor = torch.sparse_coo_tensor(indices, values, (2, 2), dtype=torch.float32)
 
@@ -84,8 +92,8 @@ def test_dropout_preserves_dtype():
 
 def test_dropout_with_fill_value_zero(mock_indices):
     """Test dropout with fill_value=0.0 (default behavior)."""
-    values = torch.tensor([5.0, 10.0, 15.0])
-    sparse_tensor = torch.sparse_coo_tensor(mock_indices, values, (3, 3))
+    values = torch.tensor([5.0, 10.0, 15.0], dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(mock_indices, values, (3, 3), dtype=values.dtype)
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5, fill_value=0.0)
 
@@ -98,8 +106,8 @@ def test_dropout_with_fill_value_zero(mock_indices):
 
 def test_dropout_with_nonzero_fill_value(mock_indices):
     """Test dropout with a non-zero fill_value."""
-    values = torch.tensor([5.0, 10.0, 15.0])
-    sparse_tensor = torch.sparse_coo_tensor(mock_indices, values, (3, 3))
+    values = torch.tensor([5.0, 10.0, 15.0], dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(mock_indices, values, (3, 3), dtype=values.dtype)
     fill_value = 99.0
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5, fill_value=fill_value)
@@ -113,9 +121,9 @@ def test_dropout_with_nonzero_fill_value(mock_indices):
 
 def test_dropout_with_negative_values():
     """Test dropout with negative values in the sparse tensor."""
-    indices = torch.tensor([[0, 1, 2], [0, 1, 2]])
-    values = torch.tensor([-1.0, -5.0, -10.0])
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (3, 3))
+    indices = torch.tensor([[0, 1, 2], [0, 1, 2]], dtype=torch.long)
+    values = torch.tensor([-1.0, -5.0, -10.0], dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(indices, values, (3, 3), dtype=values.dtype)
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5)
 
@@ -128,9 +136,11 @@ def test_dropout_preserves_cpu_device():
     """Test that dropout preserves the device."""
     device = torch.device("cpu")
 
-    indices = torch.tensor([[0, 1], [0, 1]], device=device)
-    values = torch.tensor([1.0, 2.0], device=device)
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (2, 2), device=device)
+    indices = torch.tensor([[0, 1], [0, 1]], device=device, dtype=torch.long)
+    values = torch.tensor([1.0, 2.0], device=device, dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(
+        indices, values, (2, 2), device=device, dtype=values.dtype
+    )
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5)
 
@@ -142,9 +152,11 @@ def test_dropout_preserves_cuda_device():
     """Test that dropout preserves the device."""
     device = torch.device("cuda")
 
-    indices = torch.tensor([[0, 1], [0, 1]], device=device)
-    values = torch.tensor([1.0, 2.0], device=device)
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (2, 2), device=device)
+    indices = torch.tensor([[0, 1], [0, 1]], device=device, dtype=torch.long)
+    values = torch.tensor([1.0, 2.0], device=device, dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(
+        indices, values, (2, 2), device=device, dtype=values.dtype
+    )
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5)
 
@@ -156,9 +168,11 @@ def test_dropout_preserves_mps_device():
     """Test that dropout preserves the device."""
     device = torch.device("mps")
 
-    indices = torch.tensor([[0, 1], [0, 1]], device=device)
-    values = torch.tensor([1.0, 2.0], device=device)
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (2, 2), device=device)
+    indices = torch.tensor([[0, 1], [0, 1]], device=device, dtype=torch.long)
+    values = torch.tensor([1.0, 2.0], device=device, dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(
+        indices, values, (2, 2), device=device, dtype=values.dtype
+    )
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5)
 
@@ -167,23 +181,25 @@ def test_dropout_preserves_mps_device():
 
 def test_dropout_fill_value_with_full_dropout():
     """Test that fill_value is applied correctly when dropout is 1.0."""
-    indices = torch.tensor([[0, 1], [0, 1]])
-    values = torch.tensor([1.0, 2.0])
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (2, 2))
+    indices = torch.tensor([[0, 1], [0, 1]], dtype=torch.long)
+    values = torch.tensor([1.0, 2.0], dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(indices, values, (2, 2), dtype=values.dtype)
     fill_value = 7.0
 
     result = sparse_dropout(sparse_tensor, dropout_prob=1.0, fill_value=fill_value)
 
     # All values should be the fill_value
-    assert torch.allclose(result.coalesce().values(), torch.full_like(values, fill_value))
+    assert torch.allclose(
+        result.coalesce().values(), torch.full_like(values, fill_value, dtype=values.dtype)
+    )
 
 
 def test_dropout_with_unsorted_indices():
     """Test that dropout handles unsorted indices correctly."""
     # Create a sparse tensor with unsorted/duplicate indices
-    indices = torch.tensor([[2, 0, 1, 0], [2, 0, 1, 0]])
-    values = torch.tensor([3.0, 1.0, 2.0, 4.0])
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (3, 3))
+    indices = torch.tensor([[2, 0, 1, 0], [2, 0, 1, 0]], dtype=torch.long)
+    values = torch.tensor([3.0, 1.0, 2.0, 4.0], dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(indices, values, (3, 3), dtype=values.dtype)
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5)
 
@@ -199,9 +215,9 @@ def test_dropout_with_unsorted_indices():
 
 def test_dropout_single_element():
     """Test dropout on a sparse tensor with a single element."""
-    indices = torch.tensor([[0], [0]])
-    values = torch.tensor([42.0])
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (1, 1))
+    indices = torch.tensor([[0], [0]], dtype=torch.long)
+    values = torch.tensor([42.0], dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(indices, values, (1, 1), dtype=values.dtype)
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.5)
 
@@ -216,12 +232,12 @@ def test_dropout_large_sparse_matrix():
     """Test dropout on a large sparse matrix."""
     size = 1000
     num_nonzero_elements = 500
-    rows = torch.randint(0, size, (num_nonzero_elements,))
-    cols = torch.randint(0, size, (num_nonzero_elements,))
+    rows = torch.randint(0, size, (num_nonzero_elements,), dtype=torch.long)
+    cols = torch.randint(0, size, (num_nonzero_elements,), dtype=torch.long)
 
     indices = torch.stack([rows, cols])
-    values = torch.randn(num_nonzero_elements)
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (size, size))
+    values = torch.randn(num_nonzero_elements, dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(indices, values, (size, size), dtype=values.dtype)
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.2)
 
@@ -230,7 +246,9 @@ def test_dropout_large_sparse_matrix():
 
 def test_dropout_returns_new_tensor(mock_indices, mock_values):
     """Test that dropout returns a new tensor, not a reference to the original."""
-    sparse_tensor = torch.sparse_coo_tensor(mock_indices, mock_values, (2, 2))
+    sparse_tensor = torch.sparse_coo_tensor(
+        mock_indices, mock_values, (2, 2), dtype=mock_values.dtype
+    )
 
     result = sparse_dropout(sparse_tensor, dropout_prob=0.0)
 
@@ -243,9 +261,11 @@ def test_dropout_statistical_property_moderate_rate():
     # Create a larger sparse tensor for statistical testing
     num_elements = 1000
 
-    indices = torch.tensor([list(range(num_elements)), list(range(num_elements))])
-    values = torch.ones(num_elements)
-    sparse_tensor = torch.sparse_coo_tensor(indices, values, (num_elements, num_elements))
+    indices = torch.tensor([list(range(num_elements)), list(range(num_elements))], dtype=torch.long)
+    values = torch.ones(num_elements, dtype=torch.float)
+    sparse_tensor = torch.sparse_coo_tensor(
+        indices, values, (num_elements, num_elements), dtype=values.dtype
+    )
 
     dropout_prob = 0.3
     keep_prob = 1 - dropout_prob  # Keep ~70% of elements
@@ -254,7 +274,7 @@ def test_dropout_statistical_property_moderate_rate():
     result_values = result.coalesce().values()
 
     # Count non-zero values (kept elements)
-    kept_count = (result_values != 0).sum().item()
+    kept_count = (result_values != 0).sum(dtype=torch.long).item()
     actual_keep_prob = kept_count / num_elements
 
     # Allow 10% tolerance for statistical variance
